+++ zabbix_auto_config/log.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import logging
 import sys
 
-import multiprocessing_logging
 import structlog
 
 from zabbix_auto_config.models import Settings
@@ -13,14 +12,14 @@ if sys.stderr.isatty():
 
 
 def configure_logging(config: Settings) -> None:
-    multiprocessing_logging.install_mp_handler()
-
     # Create the root logger
     root_logger = logging.getLogger()
     root_logger.setLevel(config.zac.logging.level)
 
     # Remove any existing handlers
     root_logger.handlers = []
+
+    # multiprocessing_logging.install_mp_handler()
 
     shared_processors = [
         structlog.stdlib.filter_by_level,
@@ -39,7 +38,7 @@ def configure_logging(config: Settings) -> None:
     # Configure structlog to output JSON to file and formatted text to stdout
     structlog.configure(
         processors=shared_processors,
-        # wrapper_class=structlog.stdlib.BoundLogger,
+        wrapper_class=structlog.stdlib.BoundLogger,
         context_class=dict,
         logger_factory=structlog.stdlib.LoggerFactory(),
         cache_logger_on_first_use=False,
@@ -48,8 +47,8 @@ def configure_logging(config: Settings) -> None:
     # Set up formatters and handlers
     # 1. JSON file handler
     if config.zac.logging.file:
-        file_handler = logging.FileHandler("app.log")
-        file_handler.setLevel(logging.INFO)
+        file_handler = logging.FileHandler(config.zac.logging.file)
+        file_handler.setLevel(config.zac.logging.level)
         file_formatter = structlog.stdlib.ProcessorFormatter(
             processor=structlog.processors.JSONRenderer(),
             foreign_pre_chain=shared_processors,
@@ -61,7 +60,7 @@ def configure_logging(config: Settings) -> None:
     # Human-readable formatter for console
     if config.zac.logging.stderr:
         console_handler = logging.StreamHandler(sys.stderr)
-        console_handler.setLevel(logging.INFO)
+        console_handler.setLevel(config.zac.logging.level)
         console_formatter = structlog.stdlib.ProcessorFormatter(
             processor=structlog.dev.ConsoleRenderer(colors=True),
             foreign_pre_chain=shared_processors,
@@ -73,3 +72,11 @@ def configure_logging(config: Settings) -> None:
     logging.getLogger().setLevel(config.zac.log_level)
     logging.getLogger("httpcore").setLevel(logging.ERROR)
     logging.getLogger("httpx").setLevel(logging.ERROR)
+
+    # Display the logging configuration if structured file logging is enabled
+    if config.zac.logging.file:
+        structlog.get_logger().info(
+            "Logging to file",
+            file=str(config.zac.logging.file),
+            level=config.zac.logging.level,
+        )